requirements.txt
numpy==1.26.4
opencv-python-headless
Pillow==10.2.0
streamlit==1.31.1
tensorflow==2.15.0
# tensorflow_macos==2.15.0
